division of test and non test global filter deployed via separate configurations
