+++ server/cart-service/src/test/java/com/fnd/games_store/cart/test/configuration/TestSecurityConfiguration.java
@@ -2,15 +2,17 @@ package com.fnd.games_store.cart.test.configuration;
 
 
 import com.fnd.games_store.cart.rest.UserValidationClient;
-import com.fnd.games_store.cart.test.TestJwtFilter;
+import com.fnd.games_store.cart.test.filter.TestJwtFilter;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
 
 @Configuration
-public class TestConfiguration {
+//@Profile("test")
+public class TestSecurityConfiguration {
 
     @Bean
-    public TestJwtFilter globalFilter(UserValidationClient userValidator){
+    public TestJwtFilter testGlobalFilter(UserValidationClient userValidator){
         return new TestJwtFilter(userValidator);
     }
