+++ server/cart-service/src/test/java/com/fnd/games_store/cart/test/controller_integration_test/Controller_GetCartContentTest.java
@@ -1,14 +1,15 @@
 package com.fnd.games_store.cart.test.controller_integration_test;
 
-import com.fnd.games_store.cart.CartApplication;
 import com.fnd.games_store.cart.dto.GameResponseDTO;
 import com.fnd.games_store.cart.dto.ValidationResponseDTO;
+import com.fnd.games_store.cart.test.configuration.TestSecurityConfiguration;
 import com.fnd.games_store.cart.test.utilities.ControllerTestUtilities;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.context.annotation.Import;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.test.web.servlet.MvcResult;
@@ -22,7 +23,8 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-@SpringBootTest(classes = CartApplication.class)
+@SpringBootTest
+@Import(TestSecurityConfiguration.class)
 @AutoConfigureMockMvc
 @Slf4j
 public class Controller_GetCartContentTest extends ControllerTestUtilities {
@@ -33,6 +35,7 @@ public class Controller_GetCartContentTest extends ControllerTestUtilities {
     void getCartContent_ShouldReturnProperData() throws Exception {
 
         MvcResult result = this.mvc.perform(post("/v1/content")
+                                .header("authorization", "")
                                 .content(jsonGameRequestDTO(createAppropriateGameRequestDTO(userId)))
                                 .contentType(MediaType.APPLICATION_JSON).accept(MediaType.APPLICATION_JSON))
                                 .andExpect(status().isOk())
@@ -51,6 +54,7 @@ public class Controller_GetCartContentTest extends ControllerTestUtilities {
     void getCartContent_ShouldReturnInstanceOfList() throws Exception {
 
         MvcResult result = this.mvc.perform(post("/v1/content")
+                        .header("authorization", "")
                         .content(jsonGameRequestDTO(createAppropriateGameRequestDTO(userId)))
                         .contentType(MediaType.APPLICATION_JSON).accept(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk())
