+++ server/games-service/src/main/java/com/fnd/games_store/games/configurations/WebSecurityConfiguration.java
@@ -5,6 +5,7 @@ import com.fnd.games_store.games.filter.StaffAuthorityValidationFilter;
 import org.springframework.boot.web.servlet.FilterRegistrationBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityCustomizer;
 import org.springframework.security.config.http.SessionCreationPolicy;
